requirements.txt
scikit-learn==1.4.0
xgboost==2.0.3
torch==2.2.0
numpy==1.26.4
pandas==2.2.1
joblib==1.4.0
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
redis==5.0.3
anthropic==0.25.0
reportlab==4.1.0
python-dotenv==1.0.1
imbalanced-learn==0.12.0
matplotlib==3.8.4
seaborn==0.13.2
